+++ Satori.AppServices.Tests/TimerTests/RestartTimerTests.cs
@@ -6,10 +6,10 @@ using Satori.AppServices.Services;
 using Satori.AppServices.Tests.TestDoubles;
 using Satori.AppServices.Tests.TestDoubles.AlertServices;
 using Satori.AppServices.Tests.TestDoubles.AzureDevOps;
-using Satori.AppServices.Tests.TestDoubles.AzureDevOps.Services;
 using Satori.AppServices.Tests.TestDoubles.Kimai;
 using Satori.AppServices.ViewModels;
 using Satori.Kimai.Models;
+using Satori.TimeServices;
 using Shouldly;
 
 namespace Satori.AppServices.Tests.TimerTests;
